refactor!: migrate to standard Iceberg metadata format
  Drop support for legacy Hadoop-style metadata (v*.metadata.json + version-hint.text).
  Now exclusively uses standard Iceberg format: <version>-<uuid>.metadata.json

  - Use iceberg::MetadataLocation for metadata path generation
  - Simplify extract_version_from_path to only parse standard format
  - Update doctor command to error on non-standard metadata format
  - Centralize metadata utilities in core::utils::iceberg
  - Remove version-hint.text reading/writing throughout codebase

  BREAKING CHANGE: Tables using legacy Hadoop metadata format are no longer supported.
  Migrate by renaming v<N>.metadata.json to 000<N>-<uuid>.metadata.json and removing
  version-hint.text.

+++ src/cli/commands/analyze.rs
@@ -163,7 +163,7 @@ impl AnalyzeCommand {
         };
         table.add_row(vec![
             Cell::new("Manifests"),
-            Cell::new(format_count(manifest.total_manifests as usize))
+            Cell::new(format_count(manifest.total_manifests))
                 .set_alignment(CellAlignment::Right),
             Cell::new("-").set_alignment(CellAlignment::Right),
             Cell::new(manifest_status).set_alignment(CellAlignment::Center),
